Fix bug when rendering a list of tasks
Steps to reproduce:

1. Add 3 tasks (for e.g. a, b, c)
2. Mark b as completed
3. Delete b

Now you'd see c is rendered with its checkbox checked. This happens
because b's checkbox is reused for c's.

The fix is to key each list item so that the virtual dom knows not
to reuse b's checkbox.

See https://github.com/evancz/elm-todomvc/issues/1 for a similar
issue.